+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Fix
 - **goimpl:** added buffer target of current
 - **goimpl:** mapping is now localized to buffer filetype go
+- **on_attach:** fix diagnostic and telescope diagnostic
 
 ### Refactor
 - **goimpl:** moved autocmd to on-attach